NOOM-2 Create neuron, layer and net abstractions
A artificial neural net is composed of layers of neurons that are arranged
as a graph, with each node of a layer being connected to all the nodes of
the previous and next layer. Each neuron is 'fired' when its input is high
enough (after the signoid function). It will then propigate this value to
the next nodes.

This commit creates the Neuron class, Layer Class, and a NeuralNet class.
The NeuralNet class constucts the layers and manages the running of each
layer. Each layer fires all neurons to calculate their output. The neurons
are responsible for reading all their input and deciding whether the sum
of that value is enough to 'activate' the neuron.

This is not finished and does not take the synapse (edge of graph) weight
value into account, nor is there any way to train the net. Tests have not
been included yet, but will follow with the next commit.