NOOM-2: Use weight to modify neuron input.
Previously, the raw sum of a neurons inputs were used to activate a neuron.
This meant that the neurons would always output the same result and would
not be able to 'learn' patterns; this is because there was no way to adjust
the output of neurons in the previous layer. Now, each neuron input value is
multiplied by its associated weight. This work will allow for
back-propogation learning by modifing the weights after observing the
result of training data.

Also
* Moved NeuralNetError into its own file and fixed syntax error.